Used a shell implementation of realpath in place of readlink and my own func.
Found a cross platform implementation of realpath and readlink by
Michael Kropat from the github project here:
https://github.com/mkropat/sh-realpath

The realpath function returns the absolute path to the actual file on
disk pointed to by a link, so I just replace the getLinkAbsoluteFilename
function I wrote with it.

Also used it to add a feature where the platform and OS specific dirs
can add links that reference the common shell env scripts in the home
folder with a link and it will get installed linked to the original file.