Fix error in _bash_history_sync3
The search and replace line in ed was previously using ^ to indicate the
previous line, but that appears to be invalid with the version of ed
I've tested with on MacOS and Linux. Replaced it with - which is
supposed to be the same thing, not sure why ^ doesn't work.

Also documented what the ed script was doing.